YEYlibrary

+++ src/main/java/com/example/library00/book/domain/Admin.java
@@ -1,16 +1,16 @@
 package com.example.library00.book.domain;
 
 public class Admin {
-    private int adminId;
+    private long adminId;
     private String password;
 
-    public void setAdminId(int adminId) {
+    public void setAdminId(long adminId) {
         this.adminId = adminId;
     }
 
     public void setPassword(String password){this.password = password;}
 
-    public int getAdminId() {
+    public long getAdminId() {
         return adminId;
     }
